Reorganização da pasta ./src/components/. A intenção é deixar a estrutura desse diretório mais parecido com a divisão dos conceitos básicos encontrados no site oficial da documentação do React, de onde os códigos foram retirados.
